Solution of task 4
Implementarion of program which allows teachers to create a multiple choice test in a class called Testpaper and to be also able to assign a minimum pass mark. The testpaper's subject should also be included. The attributes are in the following order:

1. subject
2. markscheme
3. pass_mark
As well as that, we need to create student objects to take the test itself! Create another class called Student and do the following:
- Create an attribute called test_taken and set the default as  'No tests taken'.
- Make a method called take_test(), which takes in the testpaper object they are completing and the student's answers. Compare what they wrote to the mark scheme, and append to the/create a dictionary assigned to tests_taken in the way as shown in the point below.
- Each key in the dictionary should be the testpaper subject and each value should be a string in the format seen in the examples below (whether or not the student has failed, and their percentage in brackets).