Solution of task 4
Task description:
Imagine we have a washing machine which can wash the clothes, rinse the clothes and spin the clothes but all the tasks separately. We need a system that can automate the whole task without the disturbance or interference of us. 
To solve the above-described problem, we would like to hire the Facade Method.