Solution of task 3
Task description :
Imagine you are creating an application that shows the data about all different types of vehicles present. It takes the data from APIs of different vehicle organizations in XML format and then displays the information.
But suppose at some time you want to upgrade your application with a Machine Learning algorithms that work beautifully on the data and fetch the important data only. But there is a problem, it takes data in JSON format only.
It will be a really poor approach to make changes in Machine Learning Algorithm so that it will take data in XML format.

For solving the problem we defined above, you can use Adapter Method that helps by creating an Adapter object.
To use an adapter in your code:
Client should make a request to the adapter by calling a method on it using the target interface.
Using the Adaptee interface, the Adapter should translate that request on the adaptee.
Result of the call is received the client and he/she is unaware of the presence of the Adapter’s presence.